Use unordered_map instead of vector for primitives
Technically this should be faster than a linear search + it allows overwriting primitives that you already have. For example:
```cpp
wac->interpreter->register_primitive("chip_digital_write", [](Module *m) -> bool {
    pop_args(2);
    printf("chip_digital_write\n");
    return true;
}, &twoToNoneU32);
```
